Birdview works now!
I made a stupid mistake, which was keeping it from finding the birdview keypoints
I fixed that, which is fantastic.

Beyond that some minor adjustments to how latency/duration are calculated
This is working quite well now.

Next Steps:
Analyze the data
Rename Sandbox2 and add it as the figure place

Once that's done, revisit posture space